Add systemd socket activation support
With this the socket that torch will be listening on is opened before
torch is started and passed in.  This is more flexible because the
socket no longer needs to be SOCK_INET.  We'll be using a unix domain
socket.  It also simplifies startup as I can choose to bind mount this
unix socket into containers before torch has started.

To test this out run:

    systemd-socket-activate -l 127.0.0.1:7654 python -m torch

and torch will be listening on http://localhost:7654 .

Systemd unit files are provided for reference.